change to return datas , move WdTemplate to gtypes
