+++ wdlang/src/gramma.rs
@@ -6,12 +6,10 @@ use std::{
   }
 };
 
-use crate::router::WdTemplate;
 use crate::lexer;
+pub mod gtypes;
 
-mod gtypes;
-
-pub fn main(path : PathBuf, template : WdTemplate){
+pub fn main(path : PathBuf, base_fnc : &PathBuf) -> gtypes::WdTemplate{
   // path ??(/home/user/project/Pages/__init__.wd ) // encoded utf-8 (ascii)
   // from path read line 
   // get tokens from line 
@@ -19,20 +17,22 @@ pub fn main(path : PathBuf, template : WdTemplate){
   // create sub scopes from scopes
   // create elements from scopes and sub scopes
 
+  let templ : gtypes::WdTemplate = gtypes::WdTemplate::new(&path, base_fnc);
   let scopes: Vec<String> = Vec::new();
 
-  let file = File::open(&path).expect(&format!("error on read : {}", path.display()));
-  let buffer = BufReader::new(&file);
+  let file : File = File::open(&path).expect(&format!("error on read : {}", path.display()));
+  let buffer : BufReader<&File> = BufReader::new(&file);
 
   for (index, line) in buffer.lines().enumerate(){
 
     // need create comment scope
-    let text = line.unwrap();
+    let text : String = line.unwrap();
     
-    let token = lexer::main(text, index);
+    let token : Option<lexer::ltypes::Token> = lexer::main(text, index);
     // println!("my tokens is  {:?}", token);
 
   }
 
+  return templ;
 
 }
