+++ wdlang/src/lexer.rs
@@ -1,6 +1,6 @@
 use regex::Regex;
 
-mod ltypes;
+pub mod ltypes;
 mod simbolys;
 mod tokens;
